requirements.txt
conda==4.10.1
cvxpy==1.1.17
joblib==1.2.0
numba==0.53.1
numpy==1.21.6
pandas==1.1.5
pathlib==1.0.1
Pillow==9.0.1
protobuf==3.20.3
pytz==2022.7.1
scikit-image==0.18.1
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
tensorboard==2.5.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==1.2.0
torchvision==0.9.1+cu111
tqdm==4.61.1